Refactor EvalPythonExec.
